requirements.txt
aiofiles==22.1.0
aiosmtplib==2.0.1
anyio==3.6.2
autopep8==2.0.1
bcrypt==4.0.1
black==22.12.0
blinker==1.5
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.89.1
fastapi-mail==1.2.4
greenlet==2.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.4
python-decouple==3.7
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
